+++ C4InterFlow/Structures/Interfaces/IStructuresResolver.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using C4InterFlow.Commons;
 
 namespace C4InterFlow.Structures.Interfaces
 {
@@ -12,6 +13,6 @@ namespace C4InterFlow.Structures.Interfaces
         public T? GetInstance<T>(string? alias) where T : Structure;
         public IEnumerable<string> ResolveStructures(IEnumerable<string> structures);
         public IEnumerable<Interface> GetAllInterfaces();
-        public void Validate(out IEnumerable<ValidationError> errors);
+        public void Validate(out IEnumerable<LogMessage> errors);
     }
 }
